Add handoff support for code blob views (#2622)
* Add handoff support for code blob views

* Github URL building w/ tests coverage
